Refactoring search detail into a utility file
